+++ src/languages/_english.json
@@ -362,7 +362,7 @@
     "discover": {
       "dpi": {
         "title": "DeFi Pulse Index",
-        "body": "All the top decentralized finance tokens in one. Track the industry.",
+        "body": "All the top decentralized finance\ntokens in one. Track the industry.",
         "view": "View"
       },
       "ens_create_profile": {
