+++ src/languages/_french.json
@@ -361,7 +361,7 @@
     "discover": {
       "dpi": {
         "title": "DeFi Pulse Index :)",
-        "body": "All the top decentralized finance tokens in one. Track the industry. :)",
+        "body": "All the top decentralized finance\ntokens in one. Track the industry. :)",
         "view": "View :)"
       },
       "ens_create_profile": {
